+++ c/jay/open_jay.cc
@@ -59,7 +59,7 @@ DataTable* DataTable::open_jay(const std::string& path)
   size_t i = 0;
   for (const jay::Column* jcol : *msg_columns) {
     Column* col = column_from_jay(jcol, mbuf);
-    if (col->nrows != static_cast<int64_t>(nrows)) {
+    if (col->nrows != nrows) {
       throw IOError() << "Length of column " << i << " is " << col->nrows
           << ", however the Frame contains " << nrows << " rows";
     }
@@ -69,7 +69,7 @@ DataTable* DataTable::open_jay(const std::string& path)
   }
 
   auto dt = new DataTable(std::move(columns), colnames);
-  dt->nkeys = frame->nkeys();
+  dt->nkeys = static_cast<size_t>(frame->nkeys());
   return dt;
 }
